Move stream code into its own higher level internal package.
otel fix
