+++ internal/stream/stream.go
@@ -12,7 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package storage
+// Package stream provides support for streaming contiguous entries from logs.
+package stream
 
 import (
 	"context"
