+++ storage/gcp/antispam/gcp.go
@@ -31,7 +31,7 @@ import (
 	"cloud.google.com/go/spanner/apiv1/spannerpb"
 
 	tessera "github.com/transparency-dev/trillian-tessera"
-	storage "github.com/transparency-dev/trillian-tessera/storage/internal"
+	"github.com/transparency-dev/trillian-tessera/internal/stream"
 	"google.golang.org/grpc/codes"
 	"k8s.io/klog/v2"
 )
@@ -221,7 +221,7 @@ func (f *follower) Follow(ctx context.Context, lr tessera.LogReader) {
 
 	t := time.NewTicker(time.Second)
 	var (
-		entryReader *storage.EntryStreamReader[[]byte]
+		entryReader *stream.EntryStreamReader[[]byte]
 		stop        func()
 
 		curEntries [][]byte
@@ -274,7 +274,7 @@ func (f *follower) Follow(ctx context.Context, lr tessera.LogReader) {
 					span.AddEvent("Start streaming entries")
 					next, st := lr.StreamEntries(ctx, followFrom)
 					stop = st
-					entryReader = storage.NewEntryStreamReader(next, f.bundleHasher)
+					entryReader = stream.NewEntryStreamReader(next, f.bundleHasher)
 				}
 
 				if curIndex == followFrom && curEntries != nil {
